pyproject.toml
[project]
name = "simple-fluid-simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "h5py>=3.15.1",
    "phiflow>=3.4.0",
]
